+++ src/main/kotlin/org/ldv/sushi/apisushi/dto/AlimentBoxDtoJson.kt
@@ -2,7 +2,7 @@ package org.ldv.sushi.apisushi.dto
 
 class AlimentBoxDtoJson (
     var nom: String,
-    var quantite: Double) {
+    var quantite: Float) {
 
     override fun toString(): String {
         return "AlimentBoxDtoJson(nom='$nom', quantite=$quantite)"
